% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mil_do_km.R
\name{mil_do_km}
\alias{mil_do_km}
\title{Konwersja mile na kilometry}
\usage{
mil_do_km(par)
}
\arguments{
\item{par}{zawieracjacy wartosci numeryczne (mile)}
}
\value{
wektor num
}
\description{
Funkcja sluzacaca do przeliczania podanej
odleglosci w milach na odleglosci w kilometrach
}
\examples{
mil_do_km(230)
mil_do_km(34.22)
mil_do_km(c(44,66,88))
}
