% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mile_do_km.R
\name{mile_do_km}
\alias{mile_do_km}
\title{Zamiana mil na kilometry}
\usage{
mile_do_km(x)
}
\arguments{
\item{x}{wektor zawierający odleglosc w milach}
}
\value{
wektor numeryczny
}
\description{
funkcja sluzaca do zamiany mil na kilometry
}
\examples{
mile_do_km(10)
mile_do_km(123)
mile_do_km(0)
mile_do_km(c(0,10,123))
}
