% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mil_do_km.R
\name{mil_do_km}
\alias{mil_do_km}
\title{Konwersja odleglosci (?)}
\usage{
mil_do_km(km_f)
}
\arguments{
\item{km_f}{Wektor zawierajacy odleglosci w milach ladowych}
}
\value{
wektor numeryczny
}
\description{
Funkcja sluzaca do konwersji mil ladowych na kilometry
}
\examples{
mil_do_km(765)
mil_do_km(98765)
mil_do_km(654)
}
